uppdate password and email

+++ src/main/java/com/main/sellplatform/controller/rest/UserRestController.java
@@ -42,13 +42,13 @@ public class UserRestController {
 
     @PreAuthorize("hasAnyAuthority('user:read')")
     @GetMapping(value = "/getUser{id}", produces = {MediaType.APPLICATION_JSON_VALUE})
-    public User getUser(@PathVariable Integer id){
+    public User getUser(@PathVariable Long id){
         return userService.getUser(id);
     }
 
     @PreAuthorize("hasAnyAuthority('user:delete')")
     @DeleteMapping(value = "/delete{id}", produces = {MediaType.APPLICATION_JSON_VALUE} )
-    public void deleteUser(@PathVariable Integer id){
+    public void deleteUser(@PathVariable Long id){
         userService.deleteUser(id);
     }
